feat: rename testbed-gw related names to dataspace
