+++ packages/af-shared/src/lib/api/endpoints/index.ts
@@ -20,6 +20,7 @@ export const USERS_API_BASE_URL = process.env.NEXT_PUBLIC_USERS_API_BASE_URL
   ? removeTrailingSlash(process.env.NEXT_PUBLIC_USERS_API_BASE_URL)
   : 'http://localhost:5001';
 
-export const JMF_SKILL_RECOMMENDATIONS_ENDPOINT = process.env.JMF_SKILL_RECOMMENDATIONS_ENDPOINT 
+export const JMF_SKILL_RECOMMENDATIONS_ENDPOINT = process.env
+  .JMF_SKILL_RECOMMENDATIONS_ENDPOINT
   ? process.env.JMF_SKILL_RECOMMENDATIONS_ENDPOINT
-  : "https://tyomarkkinatori.fi/hakupalvelu/api/1.0/skillrecommendation/skillrecommendation/";
+  : 'https://tyomarkkinatori.fi/hakupalvelu/api/1.0/skillrecommendation/skillrecommendation/';
