+++ packages/af-shared/src/lib/api/index.ts
@@ -2,9 +2,9 @@ import apiClient from './api-client';
 import * as auth from './services/auth';
 import * as codesets from './services/codesets';
 import * as company from './services/company';
+import * as dataspace from './services/dataspace';
 import * as jmf from './services/jmf';
 import * as profile from './services/profile';
-import * as testbedGw from './services/testbed-gw';
 
 const api = {
   client: apiClient,
@@ -13,7 +13,7 @@ const api = {
   codesets,
   profile,
   jmf,
-  testbedGw,
+  dataspace,
 };
 
 export default api;
