+++ packages/af-shared/src/lib/api/services/jmf.ts
@@ -1,22 +1,18 @@
-import { isExportedApplication } from '@/lib/utils';
 import type {
   JmfRecommendationsRequestPayload,
   JmfRecommendationsResponse,
 } from '@/types';
+import { isExportedApplication } from '@/lib/utils';
 import apiClient from '../api-client';
 import { TESTBED_API_BASE_URL } from '../endpoints';
 
 export async function getRecommendations(
   payload: JmfRecommendationsRequestPayload
 ): Promise<JmfRecommendationsResponse> {
-
   const url = isExportedApplication()
-  ? `${TESTBED_API_BASE_URL}/jmf/recommendations`
-  : `/api/jmf/recommendations`;
+    ? `${TESTBED_API_BASE_URL}/jmf/recommendations`
+    : `/api/jmf/recommendations`;
 
-  const { data } = await apiClient.post(
-    url,
-    payload
-  );
+  const { data } = await apiClient.post(url, payload);
   return data;
 }
