+++ packages/af-shared/src/lib/testing/mocks/handlers.ts
@@ -29,11 +29,11 @@ export const handlers = [
     (_, res, ctx) => res(ctx.json(MOCK_JOB_APPLICANT_INFO))
   ),
   rest.post(
-    'http://localhost/api/testbed-gw/Person/BasicInformation',
+    'http://localhost/api/dataspace/Person/BasicInformation',
     (_, res, ctx) => res(ctx.json(MOCK_PERSON_BASIC_INFO))
   ),
   rest.post(
-    'http://localhost/api/testbed-gw/Person/JobApplicantProfile',
+    'http://localhost/api/dataspace/Person/JobApplicantProfile',
     (_, res, ctx) => res(ctx.json(MOCK_JOB_APPLICANT_INFO))
   ),
   // codesets, add values as needed for tests
