+++ packages/af-shared/src/types/index.ts
@@ -1,6 +1,6 @@
 export * from './app';
 export * from './codesets';
 export * from './company';
+export * from './dataspace';
 export * from './jmf';
 export * from './profile';
-export * from './testbed';
